backend_main/requirements.txt
Flask
flask-restful
requests
google-generativeai
python-Levenshtein
